+++ lib/nikfemm/src/geometry/geometry_common.hpp
@@ -10,32 +10,32 @@ namespace nikfemm {
         COUNTERCLOCKWISE
     };
 
-    inline double geomDistance(Point p1, Point p2) {
+    inline float geomDistance(Point p1, Point p2) {
         return sqrt(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2));
     }
 
-    inline double geomAngle(Point a, Point b, Point c) {
+    inline float geomAngle(Point a, Point b, Point c) {
         // safe angle calculation
-        double ax = a.x - b.x;
-        double ay = a.y - b.y;
-        double bx = c.x - b.x;
-        double by = c.y - b.y;
-        double dot = ax * bx + ay * by;
-        double det = ax * by - ay * bx;
-        double angle = fabs(atan2(det, dot));
+        float ax = a.x - b.x;
+        float ay = a.y - b.y;
+        float bx = c.x - b.x;
+        float by = c.y - b.y;
+        float dot = ax * bx + ay * by;
+        float det = ax * by - ay * bx;
+        float angle = fabs(atan2(det, dot));
         return angle;
     }
 
-    inline double geomArea(Point a, Point b, Point c) {
-        double ab = geomDistance(a, b);
-        double bc = geomDistance(b, c);
-        double ac = geomDistance(a, c);
-        double s = (ab + bc + ac) / 2;
+    inline float geomArea(Point a, Point b, Point c) {
+        float ab = geomDistance(a, b);
+        float bc = geomDistance(b, c);
+        float ac = geomDistance(a, c);
+        float s = (ab + bc + ac) / 2;
         return sqrt(s * (s - ab) * (s - bc) * (s - ac));
     }
 
     inline Orientation geomOrientation(Point p1, Point p2, Point p3) {
-        double val = (p2.y - p1.y) * (p3.x - p2.x) -
+        float val = (p2.y - p1.y) * (p3.x - p2.x) -
                      (p2.x - p1.x) * (p3.y - p2.y);
 
         if (abs(val) < EPSILON) {
